Task: update go clients for ops manager and atlas
