+++ internal/store/automation.go
@@ -57,7 +57,7 @@ type CloudManagerClustersLister interface {
 func (s *Store) GetAutomationStatus(projectID string) (*opsmngr.AutomationStatus, error) {
 	switch s.service {
 	case config.CloudManagerService, config.OpsManagerService:
-		result, _, err := s.client.(*opsmngr.Client).AutomationStatus.Get(context.Background(), projectID)
+		result, _, err := s.client.(*opsmngr.Client).Automation.GetStatus(context.Background(), projectID)
 		return result, err
 	default:
 		return nil, fmt.Errorf("unsupported service: %s", s.service)
@@ -68,7 +68,7 @@ func (s *Store) GetAutomationStatus(projectID string) (*opsmngr.AutomationStatus
 func (s *Store) GetAutomationConfig(projectID string) (*opsmngr.AutomationConfig, error) {
 	switch s.service {
 	case config.CloudManagerService, config.OpsManagerService:
-		result, _, err := s.client.(*opsmngr.Client).AutomationConfig.Get(context.Background(), projectID)
+		result, _, err := s.client.(*opsmngr.Client).Automation.GetConfig(context.Background(), projectID)
 		return result, err
 	default:
 		return nil, fmt.Errorf("unsupported service: %s", s.service)
@@ -79,20 +79,9 @@ func (s *Store) GetAutomationConfig(projectID string) (*opsmngr.AutomationConfig
 func (s *Store) UpdateAutomationConfig(projectID string, automationConfig *opsmngr.AutomationConfig) error {
 	switch s.service {
 	case config.CloudManagerService, config.OpsManagerService:
-		_, err := s.client.(*opsmngr.Client).AutomationConfig.Update(context.Background(), projectID, automationConfig)
+		_, err := s.client.(*opsmngr.Client).Automation.UpdateConfig(context.Background(), projectID, automationConfig)
 		return err
 	default:
 		return fmt.Errorf("unsupported service: %s", s.service)
-	}
-}
-
-// ListAllProjectClusters encapsulate the logic to manage different cloud providers
-func (s *Store) ListAllProjectClusters() (*opsmngr.AllClustersProjects, error) {
-	switch s.service {
-	case config.OpsManagerService, config.CloudManagerService:
-		result, _, err := s.client.(*opsmngr.Client).AllClusters.List(context.Background())
-		return result, err
-	default:
-		return nil, fmt.Errorf("unsupported service: %s", s.service)
 	}
 }
