+++ internal/store/host_databases.go
@@ -31,7 +31,7 @@ type HostDatabaseLister interface {
 func (s *Store) HostDatabases(groupID, hostID string, opts *atlas.ListOptions) (*atlas.ProcessDatabasesResponse, error) {
 	switch s.service {
 	case config.OpsManagerService, config.CloudManagerService:
-		result, _, err := s.client.(*opsmngr.Client).HostDatabases.List(context.Background(), groupID, hostID, opts)
+		result, _, err := s.client.(*opsmngr.Client).Deployments.ListDatabases(context.Background(), groupID, hostID, opts)
 		return result, err
 	default:
 		return nil, fmt.Errorf("unsupported service: %s", s.service)
